Created function to make excel sheet and extended documentation
Wrote a function (create_excel) that takes an excel template (Template_CoMo_CountryData_temp.xlsx), edits it to be country-specific and then saves it as a new excel file. The name of the new excel file is returned so the file can be processed as it is now.

Documentation of the data sets cases, population, country_demog is added. mort_sever_data is standard across countries and thus already in the excel template. age_categories is not required, but I do not know how to delete mort_sever_data and age_categories from the data_CoMo.RData file.