moved excel data
the excel templates were moved to a new inst/extdata folder. The country-specific excel sheet is also saved to this folder as excel data appearantly needs to be kept separate from R data. I foresee some problems with this as the internet says that during installation of the package, folders within inst are copied into the main folder, in which case the function is directing to the wrong folder as it specifically directs to inst/extdata. If this does turn out to be a problem, there will be another commit fixing this :)